bugfix don't ignore codes 1 and 2 for any typeNum except 0 by default
Normally, pageNotFound codes 1 and 2 can be ignored. Through mktools
then the pageNotFoundHandling is aborted and TYPO3 renders the content.
That usually leads to
TYPO3 rendering the content of the next page that can be accessed. This
is usually
a login form, which is fine. With your own typeNums you will most likely
have a plugin
integrated via TypoScript. Since the code 1 and 2 are ignored, the
plugin is always rendered even if the
requested page is not accessible. Since that maybe
not desired as the plugin could render access-restricted information, we
will ignore codes 1 and 2
by default not if a separate typeNum is used.